Refactor V2 test to new file
